+++ src/cpu/two/types.h
@@ -22,36 +22,36 @@ struct affine_2d_t;
 struct trimesh_2d_t;
 
 struct point_2d_t {
-    float x, y;
+    double x, y;
 
     point_2d_t operator+=(const point_2d_t &p);
     point_2d_t operator-=(const point_2d_t &p);
     point_2d_t operator*=(const point_2d_t &p);
-    point_2d_t operator*=(float s);
+    point_2d_t operator*=(double s);
     point_2d_t operator/=(const point_2d_t &p);
-    point_2d_t operator/=(float s);
+    point_2d_t operator/=(double s);
     bool operator==(const point_2d_t &p) const;
     bool operator!=(const point_2d_t &p) const;
     bool operator<(const point_2d_t &p) const;
     point_2d_t operator<<=(const affine_2d_t &a);
 
     point_2d_t normalize() const;
-    point_2d_t rotate(float angle) const;
-    float determinant(const point_2d_t &other) const;
-    float length() const;
-    float dot(const point_2d_t &other) const;
-    float cross(const point_2d_t &other) const;
-    float angle(const point_2d_t &other) const;
+    point_2d_t rotate(double angle) const;
+    double determinant(const point_2d_t &other) const;
+    double length() const;
+    double dot(const point_2d_t &other) const;
+    double cross(const point_2d_t &other) const;
+    double angle(const point_2d_t &other) const;
 
     barycentric_coordinates_t barycentric_coordinates(
         const triangle_2d_t &t) const;
     bool is_inside_triangle(const triangle_2d_t &t) const;
     bool is_inside_bounding_box(const bounding_box_2d_t &bb) const;
 
-    float distance_to_point(const point_2d_t &other) const;
-    float distance_to_line(const line_2d_t &l) const;
-    float distance_to_triangle(const triangle_2d_t &t) const;
-    float distance_to_bounding_box(const bounding_box_2d_t &bb) const;
+    double distance_to_point(const point_2d_t &other) const;
+    double distance_to_line(const line_2d_t &l) const;
+    double distance_to_triangle(const triangle_2d_t &t) const;
+    double distance_to_bounding_box(const bounding_box_2d_t &bb) const;
 
     std::optional<point_2d_t> project_to_line(const line_2d_t &l) const;
 
@@ -61,10 +61,10 @@ struct point_2d_t {
 point_2d_t operator+(const point_2d_t &p1, const point_2d_t &p2);
 point_2d_t operator-(const point_2d_t &p1, const point_2d_t &p2);
 point_2d_t operator*(const point_2d_t &p1, const point_2d_t &p2);
-point_2d_t operator*(float s, const point_2d_t &p);
-point_2d_t operator*(const point_2d_t &p, float s);
+point_2d_t operator*(double s, const point_2d_t &p);
+point_2d_t operator*(const point_2d_t &p, double s);
 point_2d_t operator/(const point_2d_t &p1, const point_2d_t &p2);
-point_2d_t operator/(const point_2d_t &p, float s);
+point_2d_t operator/(const point_2d_t &p, double s);
 
 bool is_convex(const point_2d_t &a, const point_2d_t &b, const point_2d_t &c);
 
@@ -82,10 +82,10 @@ struct line_2d_t {
     std::vector<point_2d_t> triangle_intersection(const triangle_2d_t &t) const;
     bool intersects_bounding_box(const bounding_box_2d_t &b) const;
 
-    float distance_to_point(const point_2d_t &p) const;
-    float distance_to_line(const line_2d_t &l) const;
-    float distance_to_triangle(const triangle_2d_t &t) const;
-    float distance_to_bounding_box(const bounding_box_2d_t &bb) const;
+    double distance_to_point(const point_2d_t &p) const;
+    double distance_to_line(const line_2d_t &l) const;
+    double distance_to_triangle(const triangle_2d_t &t) const;
+    double distance_to_bounding_box(const bounding_box_2d_t &bb) const;
 
     std::string to_string() const;
 };
@@ -98,7 +98,7 @@ struct triangle_2d_t {
     bool operator<(const triangle_2d_t &p) const;
     triangle_2d_t operator<<=(const affine_2d_t &a);
 
-    float area() const;
+    double area() const;
     point_2d_t center() const;
     std::vector<point_2d_t> vertices() const;
     std::vector<line_2d_t> edges() const;
@@ -111,10 +111,10 @@ struct triangle_2d_t {
     bool intersects_bounding_box(const bounding_box_2d_t &bb) const;
     bool intersects_triangle(const triangle_2d_t &t) const;
     std::vector<point_2d_t> triangle_intersection(const triangle_2d_t &t) const;
-    float distance_to_point(const point_2d_t &p) const;
-    float distance_to_line(const line_2d_t &l) const;
-    float distance_to_triangle(const triangle_2d_t &t) const;
-    float distance_to_bounding_box(const bounding_box_2d_t &bb) const;
+    double distance_to_point(const point_2d_t &p) const;
+    double distance_to_line(const line_2d_t &l) const;
+    double distance_to_triangle(const triangle_2d_t &t) const;
+    double distance_to_bounding_box(const bounding_box_2d_t &bb) const;
     point_2d_t point_from_barycentric_coords(
         const barycentric_coordinates_t &b) const;
 
@@ -123,18 +123,18 @@ struct triangle_2d_t {
 
 struct circle_2d_t {
     point_2d_t center;
-    float radius;
+    double radius;
 
     circle_2d_t();
-    circle_2d_t(const point_2d_t &c, float r);
+    circle_2d_t(const point_2d_t &c, double r);
     circle_2d_t(const triangle_2d_t &t);
 
     bool operator==(const circle_2d_t &c) const;
     bool operator!=(const circle_2d_t &c) const;
 
-    float area() const;
-    float circumference() const;
-    bool contains_point(const point_2d_t &p, float tolerance = 0.0f) const;
+    double area() const;
+    double circumference() const;
+    bool contains_point(const point_2d_t &p, double tolerance = 0.0f) const;
 
     std::string to_string() const;
 };
@@ -153,7 +153,7 @@ struct bounding_box_2d_t {
     bool operator!=(const bounding_box_2d_t &bb) const;
     bounding_box_2d_t operator<<=(const affine_2d_t &a);
 
-    float area() const;
+    double area() const;
     point_2d_t center() const;
     std::vector<point_2d_t> vertices() const;
     std::vector<line_2d_t> edges() const;
@@ -161,10 +161,10 @@ struct bounding_box_2d_t {
     bool intersects_triangle(const triangle_2d_t &t) const;
     bool contains_point(const point_2d_t &p) const;
     bool contains_triangle(const triangle_2d_t &t) const;
-    float distance_to_point(const point_2d_t &p) const;
-    float distance_to_line(const line_2d_t &l) const;
-    float distance_to_triangle(const triangle_2d_t &t) const;
-    float distance_to_bounding_box(const bounding_box_2d_t &bb) const;
+    double distance_to_point(const point_2d_t &p) const;
+    double distance_to_line(const line_2d_t &l) const;
+    double distance_to_triangle(const triangle_2d_t &t) const;
+    double distance_to_bounding_box(const bounding_box_2d_t &bb) const;
 
     std::string to_string() const;
 };
@@ -180,7 +180,7 @@ struct polygon_2d_t {
     polygon_2d_t operator+(const polygon_2d_t &p) const;
     polygon_2d_t operator+=(const polygon_2d_t &p);
 
-    float signed_area() const;
+    double signed_area() const;
     bool is_clockwise() const;
     void reverse();
     polygon_2d_t convex_hull() const;
@@ -197,12 +197,13 @@ struct affine_2d_t {
     // r10 r11 ty
     // 0   0   1
 
-    float r00, r01, r10, r11, tx, ty;
+    double r00, r01, r10, r11, tx, ty;
 
-    affine_2d_t(float r00, float r01, float r10, float r11, float tx, float ty);
-    affine_2d_t(std::optional<float> rot = std::nullopt,
-                std::optional<std::tuple<float, float>> trans = std::nullopt,
-                std::optional<float> scale = std::nullopt);
+    affine_2d_t(double r00, double r01, double r10, double r11, double tx,
+                double ty);
+    affine_2d_t(std::optional<double> rot = std::nullopt,
+                std::optional<std::tuple<double, double>> trans = std::nullopt,
+                std::optional<double> scale = std::nullopt);
 
     affine_2d_t operator*=(const affine_2d_t &a);
